requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.9.1
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.16
filetype==1.2.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.89.0
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.11.0
google-generativeai==0.3.2
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.0
groq==0.22.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==7.2.1
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.23
langchain-community==0.0.38
langchain-core==0.3.54
langchain-google-genai==2.1.3
langchain-google-vertexai==1.0.4
langchain-groq==0.3.2
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langchain-xai==0.2.3
langsmith==0.1.147
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.4.3
mypy-extensions==1.0.0
narwhals==1.35.0
numpy==1.26.4
openai==1.75.0
orjson==3.10.16
packaging==23.2
pandas==2.2.3
pillow==10.4.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.6
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
shapely==2.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
streamlit==1.31.1
tenacity==8.5.0
textstat==0.7.3
tiktoken==0.9.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
validators==0.34.0
watchdog==6.0.0
websockets==15.0.1
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
